# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Leet-Pi API
  description: |
    Backend escalable para MVP de evaluación de ejercicios de programación en C.

    ## Autenticación
    La mayoría de los endpoints requieren un JWT Bearer token.  
    Obtené el token haciendo `POST /api/auth/login` con un token Auth0 válido en el header `Authorization: Bearer <auth0_token>`.  
    El servidor retorna el JWT de la plataforma en el header de respuesta **`X-Auth-Token`**. Usá ese valor para llamadas autenticadas subsecuentes.

    ## Roles
    - **student** – puede enviar soluciones y ver sus propias submissions.
    - **admin** – acceso a operaciones administrativas de guías, ejercicios y usuarios normales.
    - **superadmin** – nivel más alto: todo lo que puede admin más gestión de admins y asignación de roles.
  version: 1.0.0
  contact:
    name: ITBA – leet-pi team
servers:
  - url: http://localhost:3001/api
    description: Desarrollo local
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por la plataforma (`X-Auth-Token`). Obtenido vía `POST /api/auth/login`.
    Auth0Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT emitido directamente por Auth0. Requerido únicamente en `POST /api/auth/login`.

  # ─── Common response schemas ───────────────
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          description: Payload variado según el endpoint
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            statusCode:
              type: integer
          required:
            - message
            - statusCode
      required:
        - success
        - error

    # ─── Domain schemas ──────────────────────
...
    SubmissionResult:
      type: object
      properties:
        submissionId:
          type: integer
          description: Timestamp (ms) usado como ID único
        overallStatus:
          type: string
          enum: [pending, approved, failed, compilation_error]
        message:
          type: string
        score:
          type: number
        totalTests:
          type: integer
        passedTests:
          type: integer
        failedTests:
          type: integer
        compilationError:
          type: string
          nullable: true
        testResults:
          type: array
          items:
            type: object
        executionTime:
          type: number
          nullable: true
        memoryUsage:
          type: number
          nullable: true
        timestamp:
          type: string
          format: date-time
    SubmissionRecord:
      type: object
      properties:
        id:
          type: integer
        guideNumber:
          type: integer
        exerciseNumber:
          type: integer
        code:
          type: string
        success:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    GuideWithExercises:
      type: object
      properties:
        guideNumber:
          type: integer
        enabled:
          type: boolean
        exercises:
          type: array
          items:
            type: object
            properties:
              exerciseNumber:
                type: integer
              enabled:
                type: boolean

  # ─── Reusable responses ─────────────────────
  responses:
    Unauthorized:
      description: Token ausente o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: User not authenticated
              statusCode: 401
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ─────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────
paths:
...
  /submissions:
    post:
      tags: [Submissions]
      summary: Enviar solución para evaluación
      description: |
        Envía código C para ser evaluado por el juez. Verifica que el ejercicio exista, esté habilitado y la guía no haya vencido.
      operationId: submitSolution
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [exerciseNumber, guideNumber, code]
              properties:
                exerciseNumber:
                  type: integer
                  minimum: 1
                  description: Número del ejercicio
                guideNumber:
                  type: integer
                  minimum: 1
                  description: Número de la guía
                code:
                  type: string
                  minLength: 1
                  maxLength: 10000
                  description: Código fuente en C
            example:
              exerciseNumber: 3
              guideNumber: 1
              code: |
                #include <stdio.h>
                int main() {
                    printf("Hello World\n");
                    return 0;
                }
      responses:
        '200':
          description: Solución evaluada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubmissionResult'
              example:
                success: true
                message: Solution submitted successfully
                data:
                  submissionId: 1704067200000
                  overallStatus: approved
                  message: All tests passed successfully
                  score: 100
                  totalTests: 5
                  passedTests: 5
                  failedTests: 0
                  compilationError: null
                  testResults: []
                  executionTime: 0.023
                  memoryUsage: 1024
                  timestamp: '2024-01-01T00:00:00.000Z'
        '400':
          description: Request inválido (validación Joi)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Ejercicio no encontrado, deshabilitado o con deadline vencido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /submissions/exercises/available:
    get:
      tags: [Submissions]
      summary: Listar guías y ejercicios disponibles
      description: Retorna todas las guías habilitadas con sus ejercicios habilitados. No requiere autenticación.
      operationId: getAvailableExercises
      responses:
        '200':
          description: Guías disponibles
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/GuideWithExercises'
              example:
                success: true
                message: Available exercises retrieved successfully
                data:
                  - guideNumber: 1
                    enabled: true
                    exercises:
                      - exerciseNumber: 1
                        enabled: true
                      - exerciseNumber: 2
                        enabled: true
        '500':
          $ref: '#/components/responses/InternalError'
  /submissions/my:
    get:
      tags: [Submissions]
      summary: Submissions del usuario autenticado
      description: Retorna todas las submissions del usuario ordenadas por fecha descendente.
      operationId: getUserSubmissions
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de submissions
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/SubmissionRecord'
              example:
                success: true
                message: User submissions retrieved successfully
                data:
                  - id: 1704067200000
                    guideNumber: 1
                    exerciseNumber: 3
                    code: '#include <stdio.h>...'
                    success: true
                    createdAt: '2024-01-01T00:00:00.000Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /submissions/{submissionId}:
    get:
      tags: [Submissions]
      summary: Obtener submission por ID
      description: Retorna una submission específica del usuario autenticado. El `submissionId` es el timestamp (ms) retornado al crear la submission.
      operationId: getSubmissionById
      security:
        - BearerAuth: []
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: integer
          description: ID de la submission (timestamp en ms)
      responses:
        '200':
          description: Submission encontrada
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/SubmissionRecord'
        '400':
          description: ID de submission inválido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /submissions/{submissionId}/status:
    get:
      tags: [Submissions]
      summary: Estado de una submission en el juez
      description: Consulta el estado de evaluación de una submission directamente al servicio del juez. No requiere autenticación.
      operationId: getSubmissionStatus
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: integer
          description: ID de submission en el juez de código
      responses:
        '200':
          description: Estado obtenido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: Submission status retrieved successfully
                data:
                  status: completed
        '500':
          $ref: '#/components/responses/InternalError'

  # ═══════════════════════════════════════════
  # ADMIN
  # ═══════════════════════════════════════════
...